Make certauth san module pass correctly
The certauth san module should pass if it does not find any SANs in
the certificate of the types it can match against; the presence of
other types of SANs should not cause it to explicitly deny a
certificate.  Check for an empty result from
crypto_retrieve_cert_sans() in verify_client_san(), instead of
returning ENOENT from crypto_retrieve_cert_sans() when there are no
SANs at all.

ticket: 8561