Fix edge cases in KDC stash file warning
Don't warn if the stashed key has ENCTYPE_NULL.  This happens when a
database module doesn't use master keys (Sabma, kdb_test, etc.).

Don't pass the stash file name; this is NULL if it isn't explicitly
set, and the default location could depend on the KDB module.  The
administrator does not need to know the stashed key location to
migrate the master key.

ticket: 9045 (new)